yolo_shopping/requirements.txt
absl-py==0.7.0
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
Click==7.0
cycler==0.10.0
decorator==4.3.2
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.0.21
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
